Added 3 features

+++ backend/routes/postRoutes.js
@@ -4,6 +4,7 @@ const multer = require("multer");
 const router = express.Router();
 
 const Post = require('../models/post');
+const checkAuth = require("../middleware/check-auth");  
 
 const MIME_TYPE_MAP = {
     'image/png': 'png',
@@ -27,13 +28,14 @@ const storage = multer.diskStorage({
     }
 });  
 
-router.post("/", multer({storage: storage}).single("image"), (req, res, next) => {
+router.post("/", checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
     const url = req.protocol + '://' + req.get("host");
 
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
-        imagePath: url + "/images/" + req.file.filename
+        imagePath: url + "/images/" + req.file.filename,
+        creator: req.userData.userId  
     })
     post.save().then(result => {
         res.status(201).json({
@@ -43,29 +45,44 @@ router.post("/", multer({storage: storage}).single("image"), (req, res, next) =>
                 id: result._id
             }
         })
+    }).catch(error => {
+        res.status(500).json({  
+            message: "Creating A Post Failed!"  
+          });  
     })
 })
 
-// router.post("/", (req, res) => {
-//     const post = new Post({
-//         title: req.body.title,
-//         content: req.body.content
-//     });
+router.get("/", (req, res) => {
+  const pageSize = +req.query.pagesize;
+  const currentPage = +req.query.currentpage;
+  const postQuery = Post.find().populate("creator", "email");
 
-//     post.save();
-//     res.status(201).json({
-//         message: 'Post added successfully'
-//     })
-// })
+  if (pageSize && currentPage) {
+    postQuery.skip(pageSize * (currentPage - 1)).limit(pageSize);
+  }
 
-router.get("/", (_req, res) => {
-    Post.find().then(documents => {
-        res.status(200).json({
-            message: 'Posts successfully fetched',
-            posts: documents
-        })
+  let fetchedPosts;
+
+  postQuery
+    .then(docs => {
+      fetchedPosts = docs;
+      return Post.countDocuments();
     })
-})
+    .then(count => {
+      res.status(200).json({
+        message: "Posts successfully fetched",
+        posts: fetchedPosts,
+        maxPosts: count,
+      });
+    })
+    .catch(error => {
+      res.status(500).json({
+        message: "Fetching posts failed!",
+        error: error,
+      });
+    });
+});
+
 
 router.get("/:id",(req, res, next)=>{  
     Post.findById(req.params.id).then(post =>{  
@@ -74,12 +91,19 @@ router.get("/:id",(req, res, next)=>{
         }else{  
           res.status(484).json({message: 'Post not Found!'});  
         }  
-      });  
+      }).catch(error => {
+        res.status(500).json({
+            message: 'Fetching post failed!',
+            error: error
+        });
+      })  
 }); 
 
 router.put("/:id", 
+    checkAuth,
     multer({ storage: storage }).single("image"),
     (req, res, next) => {  
+        console.log(req.userData)
         let imagePath = req.body.imagePath;
         if (req.file) {
             const url = req.protocol + '://' + req.get("host");
@@ -90,22 +114,39 @@ router.put("/:id",
             _id: req.body.id,  
             title: req.body.title,  
             content: req.body.content,
-            imagePath: imagePath
+            imagePath: imagePath,
+            creator: req.userData.userId
         });  
-        Post.updateOne({_id: req.params.id}, post).then(result => { 
-            console.log(result);  
-            res.status(200).json({ message: "Update Successful!" });  
+        Post.updateOne({ _id: req.params.id,   
+            creator: req.userData.userId }, post).then(result => {  
+            if(result.modifiedCount > 0){  
+                console.log(result)
+                res.status(200).json({ message: "Update successful!" });  
+            }else{  
+                console.log(result)
+                res.status(401).json({ message: "Not Authorized!" });  
+            }  
         }).catch(error => {
-            res.status(500).json({ message: "Update failed!", error });
-        });
+            res.status(500).json({  
+                message: "Couldn't Update Post"  
+              });  
+        });  
 });
 
-router.delete("/:id", (req, res) => {
-    Post.deleteOne({_id: req.params.id }).then(result => {
-        console.log(result);
-        console.log(req.params.id)
-        res.status(200).json({message: "Post deleted"})
-    })
+router.delete("/:id", checkAuth, (req, res) => {
+    console.log(req.userData)
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {  
+        if(result.deletedCount > 0){  
+            res.status(200).json({ message: "Delete successful!" });  
+          }else{  
+            res.status(401).json({ message: "Not Authorized!" });  
+          }  
+    }).catch(error =>{  
+        res.status(500).json({  
+          message: "Deletion Not Done!" ,
+          error: error
+        });  
+      });    
 }) 
 
 module.exports = router;
